requirements.txt
PyGithub==1.50
taskw==1.2.0
